requirements.txt
streamlit==1.15.0
pytesseract==0.3.10
Pillow==9.1.1
